vk: move things around between {r,vk}_texture
Another step in preparation for new hash table and better lifetime
management
