+++ ref/vk/vk_descriptor.h
@@ -27,7 +27,7 @@ struct xvk_image_s;
 typedef union {
 	VkDescriptorBufferInfo buffer;
 	VkDescriptorImageInfo image;
-	VkDescriptorImageInfo *image_array;
+	const VkDescriptorImageInfo *image_array;
 	VkWriteDescriptorSetAccelerationStructureKHR accel;
 	const struct r_vk_image_s *image_object;
 } vk_descriptor_value_t;
