+++ ref/vk/vk_textures.h
@@ -1,4 +1,7 @@
 #pragma once
+
+#include "r_textures.h"
+
 #include "vk_core.h"
 #include "vk_image.h"
 #include "vk_const.h"
@@ -27,52 +30,18 @@ typedef struct vk_texture_s
 	// int used_maps_ago;
 } vk_texture_t;
 
-#define MAX_LIGHTMAPS 256
-#define MAX_SAMPLERS 8 // TF_NEAREST x 2 * TF_BORDER x 2 * TF_CLAMP x 2
-
-typedef struct vk_textures_global_s
-{
-	poolhandle_t mempool;
-
-	// TODO Fix these at compile time statically, akin to BLUE_NOISE_TEXTURE_ID
-	int defaultTexture;   	// use for bad textures
-	int particleTexture;
-	int whiteTexture;
-	int grayTexture;
-	int blackTexture;
-	int solidskyTexture;	// quake1 solid-sky layer
-	int alphaskyTexture;	// quake1 alpha-sky layer
-	int lightmapTextures[MAX_LIGHTMAPS];
-	int dlightTexture;	// custom dlight texture
-	int cinTexture;      	// cinematic texture
-
-// Hardcoded expected blue noise texture slot
-// TODO consider moving it into a separate resource bindable by request
-// TODO make it a 3D texture. Currently it's just a sequence of BLUE_NOISE_SIZE textures, loaded into consecutive slots.
-#define BLUE_NOISE_TEXTURE_ID 7
-
-// Hardcode blue noise texture size to 64x64x64
-#define BLUE_NOISE_SIZE 64
-
-	// TODO wire it up for ref_interface_t return
-	qboolean fCustomSkybox;
-
-	// TODO move to vk_textures/g_textures
-	vk_texture_t skybox_cube;
-
-	// All textures descriptors in their native formats used for RT
-	VkDescriptorImageInfo dii_all_textures[MAX_TEXTURES];
-} vk_textures_global_t;
-
-// TODO rename this consistently
-extern vk_textures_global_t tglob;
+typedef enum {
+	kColorspaceNative,
+	kColorspaceLinear,
+	kColorspaceGamma,
+} colorspace_hint_e;
 
 qboolean R_VkTexturesInit( void );
 void R_VkTexturesShutdown( void );
 
-// Functions only used in this renderer
-vk_texture_t *R_TextureGetByIndex( int index );
+qboolean R_VkTexturesSkyboxUpload( const char *name, rgbdata_t *const sides[6], colorspace_hint_e colorspace_hint, qboolean placeholder);
 
+qboolean R_VkTextureUpload(vk_texture_t *tex, rgbdata_t *const *const layers, int num_layers, qboolean cubemap, colorspace_hint_e colorspace_hint);
 
 // FIXME s/R_/R_Vk/
 void R_TextureAcquire( unsigned int texnum );
@@ -80,21 +49,10 @@ void R_TextureRelease( unsigned int texnum );
 
 #define R_TextureUploadFromBufferNew(name, pic, flags) R_TextureUploadFromBuffer(name, pic, flags, false)
 
-typedef enum {
-	kColorspaceNative,
-	kColorspaceLinear,
-	kColorspaceGamma,
-} colorspace_hint_e;
-
 int R_TextureUploadFromFileEx( const char *filename, colorspace_hint_e colorspace, qboolean force_reload );
-int R_TextureFindByNameF( const char *fmt, ...);
-
-// Tries to find a texture by its short name
-// Full names depend on map name, wad name, etc. This function tries them all.
-// Returns -1 if not found
-int R_TextureFindByNameLike( const char *texture_name );
-
 // Used by materials to piggy-back onto texture name-to-index hash table
 int R_TextureCreateDummy_FIXME( const char *name );
 
 VkDescriptorImageInfo R_VkTextureGetSkyboxDescriptorImageInfo( void );
+const VkDescriptorImageInfo* R_VkTexturesGetAllDescriptorsArray( void );
+VkDescriptorSet R_VkTextureGetDescriptorUnorm( uint index );
